backend/requirements.txt
flask
flask-cors
pdfminer.six
spacy
